refactor: Move trucation logic to a separate memory class
