+++ tests/test_memory_filter.py
@@ -7,7 +7,7 @@ from turbo_chat.utils.tokens import count_tokens, get_max_tokens_length
 
 @test("contains returns True when memory filter works")
 async def test_memory_filter():
-    @turbo()
+    @turbo(memory_class=LocalTruncatedMemory)
     async def example(zodiac: str, memory):
         for _ in range(50_000):
             yield System(content="You are a fortune teller")
