+++ turbo_chat/memory/__init__.py
@@ -2,7 +2,9 @@
 # This is so ruff doesn't remove * imports
 
 from .local_memory import *
+from .truncated_memory import *
 
 __all__ = [
     "LocalMemory",
+    "LocalTruncatedMemory",
 ]
