+++ turbo_chat/memory/local_memory.py
@@ -1,11 +1,7 @@
 from typing import List
 
-import pydantic
-
-
 from ..types.memory import BaseMemory
-from ..types.messages import Message, MessageDict
-from ..utils import count_tokens, get_max_tokens_length
+from ..types.messages import Message
 
 __all__ = [
     "LocalMemory",
@@ -13,7 +9,7 @@ __all__ = [
 
 
 # Abstract implementations
-class LocalMemory(BaseMemory, pydantic.BaseModel):
+class LocalMemory(BaseMemory):
     """Store messages in an in-memory list"""
 
     state: dict = {}
@@ -37,38 +33,3 @@ class LocalMemory(BaseMemory, pydantic.BaseModel):
                 **new_state,
             }
         )
-
-    async def prepare_prompt(
-        self,
-        max_tokens: int = 0,
-    ) -> List[MessageDict]:
-        """Drop messages iteratively to keep max_tokens + length < context_window."""
-        """Keep first and last message."""
-
-        max_tokens = max_tokens or 0
-        context_window: int = get_max_tokens_length(self.model) - max_tokens
-        messages = await super().prepare_prompt()
-
-        # Make sure first message is within limits
-        first, *rest = messages
-        assert count_tokens([first], self.model) <= context_window
-
-        if not len(rest):
-            return messages
-
-        # Make sure first and last combined are also within limits
-        *middle, last = rest
-        total = count_tokens([first, last], self.model)
-        assert total <= context_window
-
-        # Iteratively filter middle (remove from end)
-        final = [first]
-
-        while (
-            len(middle)
-            and (total := total + count_tokens(middle[-1:], self.model))
-            <= context_window
-        ):
-            final.insert(1, middle.pop())
-
-        return final
